Simplify RuntimeContext (#5078)
* remove SingleDataSourceRuntimeContext

* remove MultipleDataSourcesRuntimeContext

* move package

* refactor AbstractRuntimeContext's constructor
