+++ sharding-jdbc/sharding-jdbc-core/src/main/java/org/apache/shardingsphere/shardingjdbc/jdbc/core/context/impl/EncryptRuntimeContext.java
@@ -15,21 +15,30 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.shardingjdbc.jdbc.core.context;
+package org.apache.shardingsphere.shardingjdbc.jdbc.core.context.impl;
 
 import org.apache.shardingsphere.encrypt.rule.EncryptRule;
+import org.apache.shardingsphere.shardingjdbc.jdbc.core.context.AbstractRuntimeContext;
 import org.apache.shardingsphere.underlying.common.database.type.DatabaseType;
+import org.apache.shardingsphere.underlying.common.rule.BaseRule;
 
 import javax.sql.DataSource;
 import java.sql.SQLException;
+import java.util.Collection;
+import java.util.Collections;
 import java.util.Properties;
 
 /**
  * Runtime context for encrypt.
  */
-public final class EncryptRuntimeContext extends SingleDataSourceRuntimeContext<EncryptRule> {
+public final class EncryptRuntimeContext extends AbstractRuntimeContext<EncryptRule> {
     
     public EncryptRuntimeContext(final DataSource dataSource, final EncryptRule encryptRule, final Properties props, final DatabaseType databaseType) throws SQLException {
         super(dataSource, encryptRule, props, databaseType);
+    }
+    
+    @Override
+    protected Collection<BaseRule> getRules() {
+        return Collections.singletonList(getRule());
     }
 }
