+++ sharding-jdbc/sharding-jdbc-core/src/main/java/org/apache/shardingsphere/shardingjdbc/jdbc/core/context/impl/ShardingRuntimeContext.java
@@ -15,10 +15,11 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.shardingjdbc.jdbc.core.context;
+package org.apache.shardingsphere.shardingjdbc.jdbc.core.context.impl;
 
 import lombok.Getter;
 import org.apache.shardingsphere.core.rule.ShardingRule;
+import org.apache.shardingsphere.shardingjdbc.jdbc.core.context.AbstractRuntimeContext;
 import org.apache.shardingsphere.shardingjdbc.jdbc.core.datasource.metadata.CachedDatabaseMetaData;
 import org.apache.shardingsphere.transaction.ShardingTransactionManagerEngine;
 import org.apache.shardingsphere.underlying.common.database.type.DatabaseType;
@@ -35,7 +36,7 @@ import java.util.Properties;
  * Runtime context for sharding.
  */
 @Getter
-public final class ShardingRuntimeContext extends MultipleDataSourcesRuntimeContext<ShardingRule> {
+public final class ShardingRuntimeContext extends AbstractRuntimeContext<ShardingRule> {
     
     private final CachedDatabaseMetaData cachedDatabaseMetaData;
